+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Add build cache key to global
 
+# HEAD
+
+- Use webpack to build unit tests
+
 # 3.4.0 (2015-06-18)
 
 - Add app source as resolve root for modules